// src/app/model/user.ts
export class User {
    key: string;
    displayName: string;
    lowerCaseDisplayName: string;
    email: string;
    photoURL = 'https://firebasestorage.googleapis.com/v0/b/thefamily-eb9bb.appspot.com/o/images%2Fuser.png?alt=media&token=b3236d20-c28b-4022-8b81-4ff013896b64';
    uid: string;
    familyKey: string;
    admin: boolean;
}